+++ app/build.gradle
@@ -26,9 +26,10 @@ android {
     }
 
 
-    useLibrary 'org.apache.http.legacy'
+    // useLibrary 'org.apache.http.legacy'
 }
 
 dependencies {
     implementation 'com.android.support:support-v4:28.0.0'
+    implementation 'com.squareup.okhttp:okhttp:2.5.0'
 }
