+++ app/src/main/java/com/ae/apps/pnrstatus/service/status/PNRStatusService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ae.apps.pnrstatus.service;
+package com.ae.apps.pnrstatus.service.status;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -27,6 +27,7 @@ import org.json.JSONTokener;
 
 import com.ae.apps.pnrstatus.exceptions.StatusException;
 import com.ae.apps.pnrstatus.exceptions.StatusException.ErrorCodes;
+import com.ae.apps.pnrstatus.service.IStatusService;
 import com.ae.apps.pnrstatus.utils.AppConstants;
 import com.ae.apps.pnrstatus.utils.HttpUtils;
 import com.ae.apps.pnrstatus.utils.Logger;
