+++ app/src/main/res/values/strings.xml
@@ -76,5 +76,8 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.</string>
+    <string name="str_pref_dev_stub">Stub Service Response</string>
+    <string name="str_pref_dev_stub_off">Real Service</string>
+    <string name="str_pref_dev_stub_on">Stub Response</string>
 
 </resources>
