+++ app/src/main/res/xml/preferences.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android" >
 
+    <!--
     <PreferenceCategory android:title="@string/str_pref_messages" >
         <CheckBoxPreference
             android:key="pref_key_hide_past_messages"
@@ -8,6 +9,7 @@
             android:summaryOn="@string/str_pref_hide_past_summary_on"
             android:title="@string/str_pref_hide_past_messages" />
     </PreferenceCategory>
+    -->
     
     <PreferenceCategory android:title="@string/str_pref_service">
         <ListPreference 
@@ -17,6 +19,14 @@
             android:title="@string/str_pref_select_service"
             android:entries="@array/serviceNames"
             android:entryValues="@array/serviceValues"/>
+    </PreferenceCategory>
+
+    <PreferenceCategory android:title="@string/str_pref_dev_stub" >
+        <CheckBoxPreference
+            android:key="pref_key_dev_stub"
+            android:summaryOff="@string/str_pref_dev_stub_off"
+            android:summaryOn="@string/str_pref_dev_stub_on"
+            android:title="@string/str_pref_dev_stub" />
     </PreferenceCategory>
 
 </PreferenceScreen>
